feat: chage color to pink when user choose a room; componentization of choose room function
